# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sports2 API
  description: |
    REST API for the Sports2 collegiate baseball scouting and team management platform.

    ## Authentication
    All endpoints (except public ones) require:
    - `Authorization: Bearer <JWT>` header
    - `x-csrf-token` header on state-changing requests (POST, PUT, PATCH, DELETE)

    Get a CSRF token first: `GET /api/v1/auth/csrf-token`

    ## Response Format
    Success: `{ "success": true, "data": {...} }`
    Error: `{ "success": false, "error": "..." }`

    ## Scouting System
    See `docs/api/scouting-system-api.md` for the scouting system redesign including Prospects,
    20-80 grading scale, present/future grades, and recruiting pipeline.

    ## Stats & Analytics
    See `docs/api/stats-analytics-api.md` for player season/career stats, videos, leaderboards,
    and news releases (all populated by PrestoSports sync).

    ## Extended Stats & Coach's Dashboard
    See `docs/api/extended-stats-api.md` for split stats, raw Presto data, player/team game logs,
    team aggregate stats, lineup, and the coach's dashboard endpoint.
  version: 1.0.0
  contact:
    name: Sports2 Team
servers:
  - url: http://localhost:5000
    description: Local development (reference only; each deployment has its own endpoint)
security:
  - bearerAuth: []
paths:
...
  /api/v1/players:
    get:
      summary: List players
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: school_type
          in: query
          schema: { type: string, enum: [HS, COLL] }
        - name: position
          in: query
          schema: { type: string, enum: [P, C, 1B, 2B, 3B, SS, LF, CF, RF, OF, DH] }
        - name: status
          in: query
          schema: { type: string, enum: [active, inactive, graduated, transferred] }
        - name: search
          in: query
          schema: { type: string }
        - name: orderBy
          in: query
          schema: { type: string }
        - name: sortDirection
          in: query
          schema: { type: string, enum: [ASC, DESC] }
      responses:
        '200':
          description: Paginated player list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/Player' } }
                  pagination: { $ref: '#/components/schemas/Pagination' }
    post:
      summary: Create player
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/PlayerCreateInput' }
          application/json:
            schema: { $ref: '#/components/schemas/PlayerCreateInput' }
      responses:
        '201':
          description: Player created
  /api/v1/players/byId/{id}:
    get:
      summary: Get player by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Player details
        '404':
          description: Player not found
    put:
      summary: Update player
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/PlayerUpdateInput' }
          application/json:
            schema: { $ref: '#/components/schemas/PlayerUpdateInput' }
      responses:
        '200':
          description: Player updated
    delete:
      summary: Delete player
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Player deleted
  /api/v1/players/byId/{id}/stats:
    get:
      summary: Get player season & career statistics
      description: |
        Returns data from PlayerSeasonStats and PlayerCareerStats tables (populated by PrestoSports sync).
        Optionally filter to a specific season. The `current_season` field is the most recent season row.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: season
          in: query
          description: Filter to a specific season (e.g. "2025-26")
          schema: { type: string }
      responses:
        '200':
          description: Player stats with seasons and career
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      player:
                        type: object
                        properties:
                          id: { type: integer }
                          first_name: { type: string }
                          last_name: { type: string }
                          position: { type: string }
                          jersey_number: { type: integer }
                          class_year: { type: string }
                          bats: { type: string }
                          throws: { type: string }
                          photo_url: { type: string }
                      current_season: { $ref: '#/components/schemas/PlayerSeasonStats' }
                      seasons:
                        type: array
                        items: { $ref: '#/components/schemas/PlayerSeasonStats' }
                      career: { $ref: '#/components/schemas/PlayerCareerStats' }
        '404':
          description: Player not found or not on user's team
  /api/v1/players/byId/{id}/videos:
    get:
      summary: Get player videos
      description: Returns paginated videos from PlayerVideo table (populated by PrestoSports sync).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: video_type
          in: query
          description: Filter by type
          schema: { type: string, enum: [highlight, game, interview, training, promotional, other] }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 50 }
      responses:
        '200':
          description: Paginated player videos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/PlayerVideo' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        '404':
          description: Player not found or not on user's team
...
  /api/v1/players/bulk-delete:
    post:
      summary: Bulk delete players
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [player_ids]
              properties:
                player_ids: { type: array, items: { type: integer } }
      responses:
        '200':
          description: Players deleted
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    csrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
      description: Required for POST, PUT, PATCH, DELETE
  schemas:
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        pages: { type: integer }
...
    Player:
      type: object
      properties:
        id: { type: integer }
        first_name: { type: string }
        last_name: { type: string }
        position: { type: string }
        school_type: { type: string }
        school: { type: string }
        city: { type: string }
        state: { type: string }
        graduation_year: { type: integer }
        status: { type: string }
        team_id: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PlayerCreateInput:
      type: object
      required: [first_name, last_name, position]
      properties:
        first_name: { type: string }
        last_name: { type: string }
        position: { type: string, enum: [P, C, 1B, 2B, 3B, SS, LF, CF, RF, OF, DH] }
        school_type: { type: string, enum: [HS, COLL] }
        school: { type: string }
        city: { type: string }
        state: { type: string }
        graduation_year: { type: integer }
        weight: { type: integer }
        height: { type: string }
        email: { type: string }
        phone: { type: string }
    PlayerUpdateInput:
      type: object
      properties:
        first_name: { type: string }
        last_name: { type: string }
        position: { type: string }
        school_type: { type: string }
        school: { type: string }
        city: { type: string }
        state: { type: string }
        graduation_year: { type: integer }
        weight: { type: integer }
        height: { type: string }
        email: { type: string }
        phone: { type: string }
        status: { type: string }
...
    PlayerSeasonStats:
      type: object
      description: Season statistics from PrestoSports sync
      properties:
        id: { type: integer }
        player_id: { type: integer }
        team_id: { type: integer }
        season: { type: string, example: "2025-26" }
        # Batting
        games_played: { type: integer }
        games_started: { type: integer }
        at_bats: { type: integer }
        runs: { type: integer }
        hits: { type: integer }
        doubles: { type: integer }
        triples: { type: integer }
        home_runs: { type: integer }
        rbi: { type: integer }
        walks: { type: integer }
        strikeouts: { type: integer }
        stolen_bases: { type: integer }
        caught_stealing: { type: integer }
        hit_by_pitch: { type: integer }
        sacrifice_flies: { type: integer }
        sacrifice_bunts: { type: integer }
        batting_average: { type: number, format: decimal, example: 0.345 }
        on_base_percentage: { type: number, format: decimal, example: 0.410 }
        slugging_percentage: { type: number, format: decimal, example: 0.580 }
        ops: { type: number, format: decimal, example: 0.990 }
        # Pitching
        pitching_appearances: { type: integer }
        pitching_starts: { type: integer }
        innings_pitched: { type: number, format: decimal, example: 60.0 }
        pitching_wins: { type: integer }
        pitching_losses: { type: integer }
        saves: { type: integer }
        holds: { type: integer }
        hits_allowed: { type: integer }
        runs_allowed: { type: integer }
        earned_runs: { type: integer }
        walks_allowed: { type: integer }
        strikeouts_pitching: { type: integer }
        home_runs_allowed: { type: integer }
        era: { type: number, format: decimal, example: 3.50 }
        whip: { type: number, format: decimal, example: 1.20 }
        k_per_9: { type: number, format: decimal, example: 10.50 }
        bb_per_9: { type: number, format: decimal, example: 2.50 }
        # Fielding
        fielding_games: { type: integer }
        putouts: { type: integer }
        assists: { type: integer }
        errors: { type: integer }
        fielding_percentage: { type: number, format: decimal, example: 0.975 }
        # Meta
        source_system: { type: string, enum: [manual, presto] }
        last_synced_at: { type: string, format: date-time, nullable: true }
    PlayerCareerStats:
      type: object
      description: Career totals from PrestoSports sync
      properties:
        id: { type: integer }
        player_id: { type: integer }
        seasons_played: { type: integer }
        career_games: { type: integer }
        career_at_bats: { type: integer }
        career_runs: { type: integer }
        career_hits: { type: integer }
        career_doubles: { type: integer }
        career_triples: { type: integer }
        career_home_runs: { type: integer }
        career_rbi: { type: integer }
        career_walks: { type: integer }
        career_strikeouts: { type: integer }
        career_stolen_bases: { type: integer }
        career_batting_average: { type: number, format: decimal }
        career_obp: { type: number, format: decimal }
        career_slg: { type: number, format: decimal }
        career_ops: { type: number, format: decimal }
        career_pitching_appearances: { type: integer }
        career_innings_pitched: { type: number, format: decimal }
        career_wins: { type: integer }
        career_losses: { type: integer }
        career_saves: { type: integer }
        career_earned_runs: { type: integer }
        career_strikeouts_pitching: { type: integer }
        career_era: { type: number, format: decimal }
        career_whip: { type: number, format: decimal }
        source_system: { type: string, enum: [manual, presto] }
    PlayerVideo:
      type: object
      description: Player video from PrestoSports sync
      properties:
        id: { type: integer }
        player_id: { type: integer }
        team_id: { type: integer }
        title: { type: string }
        description: { type: string }
        url: { type: string }
        thumbnail_url: { type: string }
        embed_url: { type: string }
        duration: { type: integer, description: Duration in seconds }
        video_type: { type: string, enum: [highlight, game, interview, training, promotional, other] }
        provider: { type: string, description: "youtube, vimeo, hudl, etc." }
        published_at: { type: string, format: date-time }
        view_count: { type: integer }
        source_system: { type: string }
        created_at: { type: string, format: date-time }
...