# v2/src/api-client/models/Player.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sports2 API
 * REST API for the Sports2 collegiate baseball scouting and team management platform.  ## Authentication All endpoints (except public ones) require: - `Authorization: Bearer <JWT>` header - `x-csrf-token` header on state-changing requests (POST, PUT, PATCH, DELETE)  Get a CSRF token first: `GET /api/v1/auth/csrf-token`  ## Response Format Success: `{ \"success\": true, \"data\": {...} }` Error: `{ \"success\": false, \"error\": \"...\" }`  ## Scouting System See `docs/api/scouting-system-api.md` for the scouting system redesign including Prospects, 20-80 grading scale, present/future grades, and recruiting pipeline. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Player
 */
export interface Player {
    /**
     * 
     * @type {number}
     * @memberof Player
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Player
     */
    firstName?: string;
    /**
     * 
     * @type {string}
     * @memberof Player
     */
    lastName?: string;
    /**
     * 
     * @type {string}
     * @memberof Player
     */
    position?: string;
    /**
     * 
     * @type {string}
     * @memberof Player
     */
    schoolType?: string;
    /**
     * 
     * @type {string}
     * @memberof Player
     */
    school?: string;
    /**
     * 
     * @type {string}
     * @memberof Player
     */
    city?: string;
    /**
     * 
     * @type {string}
     * @memberof Player
     */
    state?: string;
    /**
     * 
     * @type {number}
     * @memberof Player
     */
    graduationYear?: number;
    /**
     * 
     * @type {string}
     * @memberof Player
     */
    status?: string;
    /**
     * 
     * @type {number}
     * @memberof Player
     */
    teamId?: number;
    /**
     * 
     * @type {Date}
     * @memberof Player
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof Player
     */
    updatedAt?: Date;
}

/**
 * Check if a given object implements the Player interface.
 */
export function instanceOfPlayer(value: object): value is Player {
    return true;
}

export function PlayerFromJSON(json: any): Player {
    return PlayerFromJSONTyped(json, false);
}

export function PlayerFromJSONTyped(json: any, ignoreDiscriminator: boolean): Player {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'firstName': json['first_name'] == null ? undefined : json['first_name'],
        'lastName': json['last_name'] == null ? undefined : json['last_name'],
        'position': json['position'] == null ? undefined : json['position'],
        'schoolType': json['school_type'] == null ? undefined : json['school_type'],
        'school': json['school'] == null ? undefined : json['school'],
        'city': json['city'] == null ? undefined : json['city'],
        'state': json['state'] == null ? undefined : json['state'],
        'graduationYear': json['graduation_year'] == null ? undefined : json['graduation_year'],
        'status': json['status'] == null ? undefined : json['status'],
        'teamId': json['team_id'] == null ? undefined : json['team_id'],
        'createdAt': json['created_at'] == null ? undefined : (new Date(json['created_at'])),
        'updatedAt': json['updated_at'] == null ? undefined : (new Date(json['updated_at'])),
    };
}

export function PlayerToJSON(json: any): Player {
    return PlayerToJSONTyped(json, false);
}

export function PlayerToJSONTyped(value?: Player | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'first_name': value['firstName'],
        'last_name': value['lastName'],
        'position': value['position'],
        'school_type': value['schoolType'],
        'school': value['school'],
        'city': value['city'],
        'state': value['state'],
        'graduation_year': value['graduationYear'],
        'status': value['status'],
        'team_id': value['teamId'],
        'created_at': value['createdAt'] == null ? value['createdAt'] : value['createdAt'].toISOString(),
        'updated_at': value['updatedAt'] == null ? value['updatedAt'] : value['updatedAt'].toISOString(),
    };
}